+++ source/WaveletDecoder/include/WaveletDecoder.h
@@ -59,7 +59,7 @@ class WaveletDecoder {
 public:
   auto decodeBand(Band &band) -> std::vector<double>;
   void transformBand(Band &band, unsigned int timescale);
-  void static decodeBlock(std::vector<double> &block_dwt, std::vector<double> &block_time,
+  void static decodeBlock(std::vector<int> &block_dwt, std::vector<double> &block_time,
                           double scalar, int dwtl);
 
 private:
